Mise à jour BDD (entity) + Dao

+++ back/src/main/java/com/csid/justread/book/infrastructure/entity/CategoryEntity.java
@@ -8,7 +8,7 @@ public class CategoryEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long Id;
+    private int Id;
 
     @Column(length = 100, nullable = false)
     private String name;
